DjangoBlog-main/requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
dj-database-url==2.1.0
Django==4.2.7
django-crispy-forms==2.1
django-environ==0.11.2
gunicorn==21.2.0
Pillow==10.1.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.8.0
whitenoise==6.6.0
tzdata==2023.3
